Feat/#31

+++ src/main/java/com/peeroreum/service/WeduService.java
@@ -4,6 +4,7 @@ import com.peeroreum.domain.Image;
 import com.peeroreum.domain.Member;
 import com.peeroreum.domain.Wedu;
 import com.peeroreum.dto.wedu.WeduDto;
+import com.peeroreum.dto.wedu.WeduReadDto;
 import com.peeroreum.dto.wedu.WeduSaveDto;
 import com.peeroreum.dto.wedu.WeduUpdateDto;
 import com.peeroreum.exception.CustomException;
@@ -13,6 +14,9 @@ import com.peeroreum.repository.MemberRepository;
 import com.peeroreum.repository.WeduRepository;
 import com.peeroreum.service.attachment.S3Service;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -33,7 +37,7 @@ public class WeduService {
         this.s3Service = s3Service;
     }
 
-    public List<WeduDto> getAllWedus() {
+    public List<WeduDto> getAll() {
         List<Wedu> weduList = weduRepository.findAll();
         List<WeduDto> weduDtoList = new ArrayList<>();
         int dDay = 0;
@@ -44,7 +48,7 @@ public class WeduService {
         return weduDtoList;
     }
 
-    public List<WeduDto> getAllMyWedus(String username) {
+    public List<WeduDto> getAllMy(String username) {
         Member member = memberRepository.findByUsername(username).orElseThrow(() -> new CustomException(ExceptionType.MEMBER_NOT_FOUND_EXCEPTION));
         Set<Wedu> myWeduList = member.getWedus();
         List<WeduDto> myWeduDtoList = new ArrayList<>();
@@ -54,58 +58,80 @@ public class WeduService {
         return myWeduDtoList;
     }
 
-    public Optional<Wedu> getWeduById(Long id) {
-        return weduRepository.findById(id);
+    public WeduReadDto getById(Long id) {
+        Wedu wedu = weduRepository.findById(id).orElseThrow(() -> new CustomException(ExceptionType.WEDU_NOT_FOUND_EXCEPTION));
+        WeduReadDto weduReadDto = WeduReadDto.builder()
+                .title(wedu.getTitle())
+                .imageUrl(wedu.getImage().getImagePath())
+                .dDay(LocalDate.now().until(wedu.getTargetDate(), ChronoUnit.DAYS))
+                .challenge(wedu.getChallenge())
+                .build();
+        return weduReadDto;
     }
 
-    public Wedu makeWedu(WeduSaveDto weduSaveDto, String username) {
+    public Wedu make(WeduSaveDto weduSaveDto, String username) {
         Member host = memberRepository.findByUsername(username).orElseThrow(()->new CustomException(ExceptionType.MEMBER_NOT_FOUND_EXCEPTION));
         Image image = imageRepository.save(s3Service.uploadImage(weduSaveDto.getFile()));
         Wedu savingWedu = Wedu.builder()
                 .title(weduSaveDto.getTitle())
-                .host(host)
                 .image(image)
+                .host(host)
                 .maximumPeople(weduSaveDto.getMaximumPeople())
-                .isSearchable(weduSaveDto.isSearchable())
                 .isLocked(weduSaveDto.isLocked())
                 .password(weduSaveDto.getPassword())
                 .grade(weduSaveDto.getGrade())
                 .subject(weduSaveDto.getSubject())
                 .gender(weduSaveDto.getGender())
+                .targetDate(weduSaveDto.getTargetDate())
                 .build();
-
+        savingWedu.addAttendant(host);
         return weduRepository.save(savingWedu);
     }
 
-    public Wedu updateWedu(Long id, WeduUpdateDto weduUpdateDto) {
+    public Wedu update(Long id, WeduUpdateDto weduUpdateDto, String username) {
+        Member member = findMember(username);
         Wedu existingWedu = weduRepository.findById(id).orElseThrow(() -> new CustomException(ExceptionType.WEDU_NOT_FOUND_EXCEPTION));
+
+        if(member != existingWedu.getHost()) {
+            throw new CustomException(ExceptionType.DO_NOT_HAVE_PERMISSION);
+        }
+
+        Image image = existingWedu.getImage();
+        if(!weduUpdateDto.getImage().isEmpty()) {
+            s3Service.deleteImage(image.getImageName());
+            imageRepository.delete(image);
+            image = imageRepository.save(s3Service.uploadImage(weduUpdateDto.getImage()));
+        }
+        existingWedu.update(image, weduUpdateDto.getMaximumPeople(), weduUpdateDto.getGender(), weduUpdateDto.isLocked(), weduUpdateDto.getPassword());
+
+        return weduRepository.save(existingWedu);
+    }
+
+    public void delete(Long id, String username) {
+        Member member = findMember(username);
+        Wedu existingWedu = weduRepository.findById(id).orElseThrow(() -> new CustomException(ExceptionType.WEDU_NOT_FOUND_EXCEPTION));
+
+        if(member != existingWedu.getHost()) {
+            throw new CustomException(ExceptionType.DO_NOT_HAVE_PERMISSION);
+        }
 
         Image image = existingWedu.getImage();
         s3Service.deleteImage(image.getImageName());
         imageRepository.delete(image);
 
-        Image newImage = imageRepository.save(s3Service.uploadImage(weduUpdateDto.getImage()));
-
-        existingWedu.setTitle(weduUpdateDto.getTitle());
-        existingWedu.setImage(newImage);
-        existingWedu.setMaximumPeople(weduUpdateDto.getMaximumPeople());
-        existingWedu.setPassword(weduUpdateDto.getPassword());
-        existingWedu.setGrade(weduUpdateDto.getGrade());
-        existingWedu.setSubject(weduUpdateDto.getSubject());
-        existingWedu.setGender(weduUpdateDto.getGender());
-
-        return weduRepository.save(existingWedu);
-    }
-    public void deleteWedu(Long id) {
-        weduRepository.deleteById(id);
+        weduRepository.delete(existingWedu);
     }
 
-    public void enrollWedu(Long id, String username) {
+    public void enroll(Long id, String username) {
         Wedu wedu = weduRepository.findById(id).orElseThrow(() -> new CustomException(ExceptionType.WEDU_NOT_FOUND_EXCEPTION));
-        Member member = memberRepository.findByUsername(username).orElseThrow(() -> new CustomException(ExceptionType.MEMBER_NOT_FOUND_EXCEPTION));
+        Member member = findMember(username);
         wedu.addAttendant(member);
         memberRepository.save(member);
         weduRepository.save(wedu);
+    }
+
+    private Member findMember(String username) {
+        return memberRepository.findByUsername(username).orElseThrow(() -> new CustomException(ExceptionType.MEMBER_NOT_FOUND_EXCEPTION));
     }
 
 }
